Move Medical orientated Tasks
Changes:
- Create new package person.health.task for any Task related to a person's health
- Move Task & MetaTasks of medical tasks
- Delete MedicalAssistence Task as it is not used
- Correct ECAOperationTest to make sure person is fit before the EVA walk
- Extend time of the harvest in the Algae Pond test
- Remove the duration modification in DigLocal task as it is not a duration based Task

Part of #1285
