+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -24,19 +24,15 @@ import com.mars_sim.core.person.ai.task.meta.ConverseMeta;
 import com.mars_sim.core.person.ai.task.meta.DelegateWorkMeta;
 import com.mars_sim.core.person.ai.task.meta.DoInventoryMeta;
 import com.mars_sim.core.person.ai.task.meta.EatDrinkMeta;
-import com.mars_sim.core.person.ai.task.meta.ExamineBodyMeta;
 import com.mars_sim.core.person.ai.task.meta.ListenToMusicMeta;
 import com.mars_sim.core.person.ai.task.meta.MeetTogetherMeta;
 import com.mars_sim.core.person.ai.task.meta.PlanMissionMeta;
 import com.mars_sim.core.person.ai.task.meta.PlayHoloGameMeta;
-import com.mars_sim.core.person.ai.task.meta.PrescribeMedicationMeta;
 import com.mars_sim.core.person.ai.task.meta.ReadMeta;
 import com.mars_sim.core.person.ai.task.meta.RecordActivityMeta;
 import com.mars_sim.core.person.ai.task.meta.RelaxMeta;
 import com.mars_sim.core.person.ai.task.meta.RepairMalfunctionMeta;
 import com.mars_sim.core.person.ai.task.meta.ReportMissionControlMeta;
-import com.mars_sim.core.person.ai.task.meta.RequestMedicalTreatmentMeta;
-import com.mars_sim.core.person.ai.task.meta.RestingMedicalRecoveryMeta;
 import com.mars_sim.core.person.ai.task.meta.ReviewJobReassignmentMeta;
 import com.mars_sim.core.person.ai.task.meta.ReviewMissionPlanMeta;
 import com.mars_sim.core.person.ai.task.meta.SalvageBuildingMeta;
@@ -44,12 +40,16 @@ import com.mars_sim.core.person.ai.task.meta.SalvageGoodMeta;
 import com.mars_sim.core.person.ai.task.meta.SelfTreatHealthProblemMeta;
 import com.mars_sim.core.person.ai.task.meta.SleepMeta;
 import com.mars_sim.core.person.ai.task.meta.TeachMeta;
-import com.mars_sim.core.person.ai.task.meta.TreatMedicalPatientMeta;
 import com.mars_sim.core.person.ai.task.meta.WorkoutMeta;
 import com.mars_sim.core.person.ai.task.meta.WriteReportMeta;
 import com.mars_sim.core.person.ai.task.meta.YogaMeta;
 import com.mars_sim.core.person.ai.task.util.MetaTask.TaskScope;
 import com.mars_sim.core.person.ai.task.util.MetaTask.WorkerType;
+import com.mars_sim.core.person.health.task.ExamineBodyMeta;
+import com.mars_sim.core.person.health.task.PrescribeMedicationMeta;
+import com.mars_sim.core.person.health.task.RequestMedicalTreatmentMeta;
+import com.mars_sim.core.person.health.task.RestingMedicalRecoveryMeta;
+import com.mars_sim.core.person.health.task.TreatMedicalPatientMeta;
 import com.mars_sim.core.robot.ai.task.ChargeMeta;
 import com.mars_sim.core.robot.ai.task.SavePowerMeta;
 import com.mars_sim.core.science.task.AssistScientificStudyResearcherMeta;
